Fix javascript RefactorRename tests and test matchers. Move matchers to common library
