+++ ycmd/tests/test_utils.py
@@ -158,3 +158,21 @@ class DummyCompleter( Completer ):
   # This method is here for testing purpose, so it can be mocked during tests
   def CandidatesList( self ):
     return []
+
+
+def LocationMatcher( filepath, line_num, column_num ):
+  return has_entries( {
+    'line_num': line_num,
+    'column_num': column_num,
+    'filepath': filepath
+  } )
+
+
+def ChunkMatcher( replacement_text, start, end ):
+  return has_entries( {
+    'replacement_text': replacement_text,
+    'range': has_entries( {
+      'start': start,
+      'end': end
+    } )
+  } )
